Simplify documentation and extract scraper for extensibility
## Documentation Consolidation (7 docs → 2 docs)

### Problem
Too many overlapping documentation files:
- ARCHITECTURE.md + ARCHITECTURE_DIAGRAM.md (redundant)
- TESTING_FLOWS.md + TEST_ENVIRONMENT.md (overlapping)
- CODE_REVIEW.md (one-time review, not living doc)

### Solution: ONE Comprehensive Guide

Created **GUIDE.md** - Single source of truth containing:
1. Quick Start
2. Architecture (with diagrams)
3. Testing (automated + flow scripts + data generation)
4. Development (standards, adding features)
5. Deployment (checklist, monitoring)
6. Troubleshooting (all scenarios)
7. Reliability Features (all 8 priorities)

**Benefits:**
✅ One file to read (vs navigating 7 files)
✅ No redundancy
✅ Easier to maintain
✅ Clear table of contents

### Archived Old Docs

Moved to `.ai/archive/` (for reference, not active):
- ARCHITECTURE.md
- ARCHITECTURE_DIAGRAM.md
- TEST_ENVIRONMENT.md
- TESTING_FLOWS.md
- CODE_REVIEW.md

### Updated References
- README.md now points to GUIDE.md
- claude.md updated to reference GUIDE.md

## Scraper Extraction for Future Extensibility

### Problem
Scraper tightly coupled, hard to add new data sources (FRED, Investing.com, etc.)

### Solution: Service Layer with Interface

Created `src/services/scrapers/`:
- **ICalendarScraper.ts** - Interface for all scrapers
  - scrape() - Returns CalendarEvent[]
  - isAvailable() - Health check
  - sourceName - Identification

- **marketwatchScraper.ts** - MarketWatch implementation
  - Class-based (was function-based)
  - Implements ICalendarScraper
  - Singleton export: marketwatchScraper
  - Legacy function export for backwards compatibility

- **src/scraper.ts** - Legacy re-export module
  - Maintains backwards compatibility
  - Deprecation notice for new code
  - Future: Will become scraperOrchestrator

### Benefits
✅ Easy to add new scrapers (implement ICalendarScraper)
✅ Backwards compatible (all existing code works)
✅ Testable in isolation
✅ Swappable implementations
✅ Future: Multi-source fallback (MarketWatch → FRED → Investing.com)

### Example Future Usage
```typescript
// src/services/scrapers/fredScraper.ts
export class FREDScraper implements ICalendarScraper {
  sourceName = 'FRED';
  async scrape() { /* ... */ }
  async isAvailable() { /* ... */ }
}

// src/services/scrapers/scraperOrchestrator.ts
const scrapers = [marketwatchScraper, fredScraper, investingComScraper];
for (const scraper of scrapers) {
  if (await scraper.isAvailable()) {
    return await scraper.scrape();
  }
}
```

## Test Results
✅ 106/106 tests passing
✅ Build successful
✅ Live scraper: "[MarketWatch] Using selector..." (class-based logging)

## Final Documentation Structure
```
README.md              - Project overview (points to GUIDE.md)
GUIDE.md               - ONE comprehensive reference (NEW)
claude.md              - AI session config
.ai/rules/             - Code standards
.ai/archive/           - Historical docs (for reference)
```

From 7 active docs → 2 active docs (README + GUIDE)
