+++ .ai/archive/ARCHITECTURE_DIAGRAM.md
